requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
numpy>=1.26.0
sounddevice>=0.4.6
faster-whisper>=1.0.0
piper-tts>=1.4.2
streamlit>=1.36.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
prometheus-client>=0.20.0
langfuse>=2.44.0
evidently>=0.4.30
pandas>=2.2.0
pytest>=8.2.0
